Renamed halotools_exceptions module to custom_exceptions
