+++ halotools/custom_exceptions.py
