+++ halotools/empirical_models/hod_designer.py
@@ -6,7 +6,7 @@ from an existing model.
 """
 
 from copy import copy 
-from ..halotools_exceptions import HalotoolsError
+from ..custom_exceptions import HalotoolsError
 from .model_factories import HodModelFactory 
 from warnings import warn
