+++ halotools/empirical_models/model_factories.py
@@ -25,7 +25,7 @@ from ..sim_manager.supported_sims import HaloCatalog
 from ..sim_manager.generate_random_sim import FakeSim
 from ..utils.array_utils import array_like_length as custom_len
 
-from ..halotools_exceptions import HalotoolsError
+from ..custom_exceptions import HalotoolsError
 from warnings import warn 
 
 @six.add_metaclass(ABCMeta)
